requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# LangChain stack (let pip resolve compatible versions)
langchain
langchain-core
langchain-community
langchain-text-splitters
langsmith

# Document Parsing
unstructured>=0.11.0
python-multipart>=0.0.6

# PDF / DOCX
pypdf>=3.17.0
python-docx>=1.1.0

# Postgres & Vector DB
psycopg2-binary>=2.9.0
pgvector>=0.2.0
sqlalchemy>=2.0.0

# LLMs
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
sentence-transformers==2.2.2

# Misc
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.1.0
requests>=2.31.0